Add functionality for script to pull repo privacy
Results now include a column/attribute called "private" that indicates
whether the repository is private or not. The PyGitHub library that the
script uses does not yet support GitHub Enterprise and therefore cannot
determine whether a repo is "internal", but "internal" repos appear to
result in private being set to true.